pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rag-system"
version = "0.1.0"
description = "A Retrieval-Augmented Generation (RAG) system built using Google's Gemini model and FAISS vector database"
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.1",
    "python-dotenv>=1.0.1",
    "google-generativeai>=0.3.2",
    "faiss-cpu>=1.7.4",
    "numpy>=1.26.4",
    "PyMuPDF>=1.23.26",
    "pydantic>=2.6.3",
    "python-multipart>=0.0.9",
    "loguru>=0.7.2",
    "langchain-google-genai>=0.0.5",
    "langchain>=0.1.12",
    "langchain-community>=0.0.27",
    "langgraph>=0.0.10",
    "typing-extensions>=4.9.0",
    "pandas>=2.2.1",
    "scikit-learn>=1.4.1",
    "requests>=2.31.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "isort>=5.13.0",
    "mypy>=1.8.0",
    "ruff>=0.1.14",
]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=app --cov-report=term-missing"

[tool.ruff]
line-length = 88
target-version = "py39"
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.isort]
known-first-party = ["app"]
